requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt setup.py
#
alabaster==0.7.10         # via sphinx
arrow==0.10.0             # via jinja2-time
babel==2.4.0              # via flask-babel, sphinx
binaryornot==0.4.3        # via cookiecutter
blinker==1.4              # via flask-mail, flask-principal, groundwork
certifi==2017.4.17        # via requests
chardet==3.0.4            # via binaryornot, requests
click==6.7                # via cookiecutter, flask, groundwork
cookiecutter==1.5.1       # via groundwork
docutils==0.13.1          # via groundwork, groundwork-web, sphinx
flask-admin==1.5.0        # via groundwork-web
flask-babel==0.11.2
flask-login==0.3.2        # via flask-security
flask-mail==0.9.1         # via flask-security
flask-principal==0.4.0    # via flask-security
flask-restless==0.17.0    # via groundwork-web
flask-security==1.7.5
flask-wtf==0.14.2         # via flask-security
flask==0.12.2             # via flask-admin, flask-babel, flask-login, flask-mail, flask-principal, flask-restless, flask-security, flask-wtf, groundwork-web
future==0.16.0            # via cookiecutter
gitdb2==2.0.2             # via gitpython
gitpython==2.1.5          # via groundwork-database
groundwork-database==0.1.4
groundwork-web==0.1.4
groundwork==0.1.14
idna==2.5                 # via requests
imagesize==0.7.1          # via sphinx
itsdangerous==0.24        # via flask, flask-security
jinja2-time==0.2.0        # via cookiecutter
jinja2==2.9.6             # via cookiecutter, flask, flask-babel, groundwork, jinja2-time, sphinx
markupsafe==1.0           # via jinja2
mimerender==0.6.0         # via flask-restless
passlib==1.7.1            # via flask-security
pathlib==1.0.1            # via groundwork
poyo==0.4.1               # via cookiecutter
pygments==2.2.0           # via sphinx
pytest-runner==2.11.1     # via groundwork-web
python-dateutil==2.6.0    # via arrow, flask-restless
python-mimeparse==1.6.0   # via mimerender
pytz==2017.2              # via babel
requests==2.17.3          # via sphinx
six==1.10.0               # via python-dateutil, sphinx
smmap2==2.0.3             # via gitdb2
snowballstemmer==1.2.1    # via sphinx
sphinx==1.6.2             # via groundwork-database, groundwork-web
sphinxcontrib-websupport==1.0.1  # via sphinx
sqlalchemy==1.1.10        # via flask-restless, groundwork-database
urllib3==1.21.1           # via requests
werkzeug==0.12.2          # via flask
whichcraft==0.4.1         # via cookiecutter
wtforms==2.1              # via flask-admin, flask-wtf
